# client_axios/model/exception-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger for MyHome
 * This is openapi specification for MyHome application.
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ExceptionDTO
 */
export interface ExceptionDTO {
    /**
     * 
     * @type {string}
     * @memberof ExceptionDTO
     */
    'details': string;
    /**
     * 
     * @type {string}
     * @memberof ExceptionDTO
     */
    'exception': string;
    /**
     * 
     * @type {string}
     * @memberof ExceptionDTO
     */
    'message': string;
    /**
     * 
     * @type {string}
     * @memberof ExceptionDTO
     */
    'timestamp': string;
}