# client_axios/model/login-parameter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger for MyHome
 * This is openapi specification for MyHome application.
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface LoginParameter
 */
export interface LoginParameter {
    /**
     * 
     * @type {string}
     * @memberof LoginParameter
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof LoginParameter
     */
    'password': string;
    /**
     * 
     * @type {boolean}
     * @memberof LoginParameter
     */
    'rememberMe': boolean;
}