# client_axios/model/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger for MyHome
 * This is openapi specification for MyHome application.
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { UserRoleEnum } from './user-role-enum';

/**
 * 
 * @export
 * @interface UserDTO
 */
export interface UserDTO {
    /**
     * 
     * @type {string}
     * @memberof UserDTO
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof UserDTO
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UserDTO
     */
    'firstName': string;
    /**
     * 
     * @type {string}
     * @memberof UserDTO
     */
    'lastName': string;
    /**
     * 
     * @type {Array<UserRoleEnum>}
     * @memberof UserDTO
     */
    'roles': Array<UserRoleEnum>;
}