# client_axios/model/user-registration-parameter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger for MyHome
 * This is openapi specification for MyHome application.
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: jeanbaptiste.wittner@outlook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UserRegistrationParameter
 */
export interface UserRegistrationParameter {
    /**
     * 
     * @type {string}
     * @memberof UserRegistrationParameter
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistrationParameter
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistrationParameter
     */
    'firstName': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistrationParameter
     */
    'lastName': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistrationParameter
     */
    'password': string;
}