# openapi.yaml
openapi: 3.0.3
info:
  title: Swagger for MyHome
  description: This is openapi specification for MyHome application.
  contact:
    email: jeanbaptiste.wittner@outlook.com
  license:
    name: GPL-3.0 License
    url: 'https://www.gnu.org/licenses/gpl-3.0.html'
  version: 0.1.0
externalDocs:
  description: Find out more about Swagger
  url: 'https://swagger.io'
servers:
  - url: 'http://localhost:6080/api'
tags:
  - name: security
    description: Management of security
paths:
#  ----- PATHS  ---------------------------------------------
#  ----- AUTHENTICATION  ------------------------------------
  /authentication/registration:
    put:
      tags:
        - security
      summary: Register a new user account
      operationId: registration
      security: []
      requestBody:
        description: Object that needs to register a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationParameter'
      responses:
        '200':
          description: successful operation
#  ----- AUTHENTICATION  ------------------------------------
components:
  schemas:
    ExceptionDTO:
      type: object
      properties:
        details:
          type: string
        exception:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - details
        - exception
        - message
        - timestamp
#  ----- PARAMETERS  ----------------------------------------
    LoginParameter:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
        rememberMe:
          type: boolean
      required:
        - username
        - password
        - rememberMe
    UserRegistrationParameter:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
          pattern: >-
            ^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
          pattern: '^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=\S+$).{8,}$'
      required:
        - username
        - email
        - firstName
        - lastName
        - password
#  ----- PARAMETERS  ----------------------------------------
#  ----- DTO  -----------------------------------------------
    UserDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRoleEnum'
      required:
        - username
        - email
        - firstName
        - lastName
        - roles
#  ----- DTO  -----------------------------------------------
#  ----- Enum -----------------------------------------------
    UserRoleEnum:
      type: string
      enum:
        - ADMIN
        - USER